requirements.txt
pandas
numpy
matplotlib
seaborn
statsmodels
scikit-learn
xgboost
lightgbm
pyyaml
dvc
joblib
mlflow==2.11.0
dagshub
streamlit
fastapi
uvicorn
httpx
requests
pytest
flake8
